% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_export.R
\name{export_drive}
\alias{export_drive}
\title{Export Google Drive folder, with conversion to locally-editable file types}
\usage{
export_drive(url, dir = ".", quiet = TRUE, progress = TRUE)
}
\arguments{
\item{url}{URL of a Google Drive folder.}

\item{dir}{Local path where subfolders and files will be downloaded.}

\item{quiet}{Logical, whether to suppress googledrive messages during each download.}

\item{progress}{Logical, whether to show a download progress bar.}
}
\value{
Nothing
}
\description{
Export Google Drive folder, with conversion to locally-editable file types
}
\examples{
export_drive()

}
